refactor(be): 엔티티 데이터 타입 수정 및 Swagger 정리 (#269)
* refactor(be): Swagger 문서 정리

* refactor(be): 주문 도메인에서 현재 부적합한 데이터 타입 변경

* OrderItem 에서 추후 값 연산 시 값의 오버플로우 방지
* 현재 설계상 데이터 값을 소수점으로 만들어서 저장하지는 않으므로 BigDecimal 대신 Long 사용

* refactor(be): 결제 도메인에서 현재 부적합한 데이터 타입 변경

* 현재 설계상 데이터 값을 소수점으로 만들어서 저장하지는 않으므로 BigDecimal 대신 Long 사용
* long 보단 Long 을 통한 표현력 강화(조금의 메모리 공간과 오버헤드로 인한 성능차이는 미미)

* refactor(be): 정산 도메인에서 현재 부적합한 데이터 타입 변경

* 현재 설계상 데이터 값을 소수점으로 만들어서 저장하지는 않으므로 BigDecimal 대신 Long 사용
* long 보단 Long 을 통한 표현력 강화(조금의 메모리 공간과 오버헤드로 인한 성능차이는 미미)
* 배달원/상점의 정산 건 수는 21억을 넘을 일이 정말 없으므로

+++ backend/src/main/java/com/deliveranything/domain/order/dto/OrderCreateRequest.java
@@ -6,7 +6,6 @@ import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
-import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderCreateRequest(
@@ -17,9 +16,9 @@ public record OrderCreateRequest(
     @NotNull Double lng,
     @Size(max = 30) String riderNote,
     @Size(max = 30) String storeNote,
-    @NotNull @Positive BigDecimal totalPrice,
-    @NotNull @Positive BigDecimal storePrice,
-    @NotNull @Positive BigDecimal deliveryPrice
+    @NotNull @Positive Long totalPrice,
+    @NotNull @Positive Long storePrice,
+    @NotNull @Positive Long deliveryPrice
 ) {
 
 }
